Extract video controls into separate file
