+++ src/plugin/contentScript.ts
@@ -1,8 +1,15 @@
 import Port = chrome.runtime.Port
 import type { VideoEvent } from '../server/VideoEvent'
 import type { MessageType } from './MessageType'
-import { acceptableTimeDifferenceBetweenClientsInSeconds } from './config'
 import { listenForBrowserActionEvents } from './contentScript/listenForBrowserActionEvents'
+import {
+  pause,
+  play,
+  setNewVideoTimeIfNecessary,
+  setupVideoEventHandlers,
+  SkipEvents,
+  skipEvents,
+} from './contentScript/videoController'
 
 const asyncSendMessage = (message: MessageType) => {
   return new Promise((resolve, reject) => {
@@ -33,7 +40,9 @@ const registerNewSession = async (): Promise<string> => {
   }
 
   const hasResult = (o: unknown): o is { result: string } => {
-    return typeof response === 'object' && response !== null && typeof (o as { result: unknown }).result === 'string'
+    return typeof response === 'object' && response !== null && typeof (
+      o as { result: unknown }
+    ).result === 'string'
   }
 
   if (hasResult(response)) {
@@ -82,118 +91,6 @@ const getOrCreateSessionID = async () => {
   }
 
   return sessionID
-}
-
-type SkipEvents = { [key in keyof HTMLMediaElementEventMap]: boolean }
-
-const setupVideoEventHandlers = (port: Port, video: HTMLVideoElement) => {
-  const playLike = ['play'] as Array<keyof HTMLMediaElementEventMap>
-  const pauseLike = ['pause'] as Array<keyof HTMLMediaElementEventMap>
-  const seekLike = ['seeking'] as Array<keyof HTMLMediaElementEventMap>
-
-  const skipEvents = [...playLike, ...pauseLike, ...seekLike].reduce((skipEvents, eventName) => (
-    {
-      ...skipEvents,
-      [eventName]: false,
-    }
-  ), {} as SkipEvents)
-
-  const registerEvent = (eventType: string, eventName: keyof SkipEvents) => {
-    const listener = () => {
-      if (!skipEvents[eventName]) {
-        console.log(`Sending ${eventName} event`)
-        const videoTime = video.currentTime
-        port.postMessage({
-          query: 'videoEvent',
-          payload: {
-            type: eventType,
-            data: { videoTime },
-          },
-        })
-      } else {
-        console.info(`Skipping ${eventName} event`)
-        skipEvents[eventName] = false
-      }
-    }
-    video?.addEventListener(eventName, listener)
-
-    return () => video?.removeEventListener(eventName, listener)
-  }
-
-  const playEventRemovers = playLike.map(eventName =>
-    registerEvent('playLikeEvent', eventName),
-  )
-
-  const pauseEventRemovers = pauseLike.map(eventName =>
-    registerEvent('pauseLikeEvent', eventName),
-  )
-
-  const seekEventRemovers = seekLike.map(eventName =>
-    registerEvent('seekLikeEvent', eventName),
-  )
-
-  const removeEventListeners = () => {
-    [...playEventRemovers, ...pauseEventRemovers, ...seekEventRemovers].forEach(remover => remover())
-  }
-
-  return { skipEvents, removeEventListeners }
-}
-
-const getDisneyPlusPlayPauseElement = () => document.querySelector<HTMLButtonElement>(
-  'div > div > div.controls__footer.display-flex > div.controls__footer__wrapper.display-flex >' +
-  ' div.controls__center > button.control-icon-btn.play-icon.play-pause-icon',
-)
-
-const play = (video: HTMLVideoElement) => {
-  if (video.paused) {
-    console.log('playing')
-    const disneyPlusPlayPauseButton = getDisneyPlusPlayPauseElement()
-    if (disneyPlusPlayPauseButton) {
-      disneyPlusPlayPauseButton.click()
-    } else {
-      video.play()
-    }
-  }
-}
-
-const pause = (video: HTMLVideoElement) => {
-  if (!video.paused) {
-    console.log('pausing')
-    const disneyPlusPlayPauseButton = getDisneyPlusPlayPauseElement()
-    if (disneyPlusPlayPauseButton) {
-      disneyPlusPlayPauseButton.click()
-    } else {
-      video.pause()
-    }
-  }
-}
-
-const setNewVideoTimeIfNecessary = (
-  video: HTMLVideoElement,
-  shouldSkipEvents: SkipEvents,
-  newVideoTime?: number,
-  force = false,
-) => {
-  if (
-    typeof newVideoTime === 'number' &&
-    (
-      force || Math.abs(video.currentTime - newVideoTime) > acceptableTimeDifferenceBetweenClientsInSeconds
-    )
-  ) {
-    shouldSkipEvents.seeking = true
-    video.currentTime = newVideoTime
-  }
-}
-
-const skipEvents = (skipEvents: SkipEvents, ...eventNames: Array<keyof SkipEvents>): void => {
-  const keys = Object.keys(skipEvents) as Array<keyof SkipEvents>
-  keys.forEach(key => {
-    skipEvents[key] = false
-  })
-
-  eventNames.forEach(eventName => {
-    skipEvents[eventName] = true
-  })
 }
 
 const onSync = (
